service: Split all existing service into interface and implementation
