+++ src/main/java/com/unconv/spring/service/BookingService.java
@@ -2,50 +2,14 @@ package com.unconv.spring.service;
 
 import com.unconv.spring.domain.Booking;
 import com.unconv.spring.model.response.PagedResult;
-import com.unconv.spring.persistence.BookingRepository;
 import java.util.Optional;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-@Service
-@Transactional
-public class BookingService {
+public interface BookingService {
+    PagedResult<Booking> findAllBookings(int pageNo, int pageSize, String sortBy, String sortDir);
 
-    private final BookingRepository bookingRepository;
+    Optional<Booking> findBookingById(Long id);
 
-    @Autowired
-    public BookingService(BookingRepository bookingRepository) {
-        this.bookingRepository = bookingRepository;
-    }
+    Booking saveBooking(Booking booking);
 
-    public PagedResult<Booking> findAllBookings(
-            int pageNo, int pageSize, String sortBy, String sortDir) {
-        Sort sort =
-                sortDir.equalsIgnoreCase(Sort.Direction.ASC.name())
-                        ? Sort.by(sortBy).ascending()
-                        : Sort.by(sortBy).descending();
-
-        // create Pageable instance
-        Pageable pageable = PageRequest.of(pageNo, pageSize, sort);
-        Page<Booking> bookingsPage = bookingRepository.findAll(pageable);
-
-        return new PagedResult<>(bookingsPage);
-    }
-
-    public Optional<Booking> findBookingById(Long id) {
-        return bookingRepository.findById(id);
-    }
-
-    public Booking saveBooking(Booking booking) {
-        return bookingRepository.save(booking);
-    }
-
-    public void deleteBookingById(Long id) {
-        bookingRepository.deleteById(id);
-    }
+    void deleteBookingById(Long id);
 }
