+++ src/main/java/com/unconv/spring/service/OfferService.java
@@ -1,32 +1,15 @@
 package com.unconv.spring.service;
 
 import com.unconv.spring.domain.Offer;
-import com.unconv.spring.persistence.OfferRepository;
 import java.util.List;
 import java.util.Optional;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-@Service
-@Transactional
-public class OfferService {
+public interface OfferService {
+    List<Offer> findAllOffers();
 
-    @Autowired private OfferRepository offerRepository;
+    Optional<Offer> findOfferById(Long id);
 
-    public List<Offer> findAllOffers() {
-        return offerRepository.findAll();
-    }
+    Offer saveOffer(Offer offer);
 
-    public Optional<Offer> findOfferById(Long id) {
-        return offerRepository.findById(id);
-    }
-
-    public Offer saveOffer(Offer offer) {
-        return offerRepository.save(offer);
-    }
-
-    public void deleteOfferById(Long id) {
-        offerRepository.deleteById(id);
-    }
+    void deleteOfferById(Long id);
 }
