+++ src/main/java/com/unconv/spring/service/RouteService.java
@@ -2,45 +2,14 @@ package com.unconv.spring.service;
 
 import com.unconv.spring.domain.Route;
 import com.unconv.spring.model.response.PagedResult;
-import com.unconv.spring.persistence.RouteRepository;
 import java.util.Optional;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-@Service
-@Transactional
-public class RouteService {
+public interface RouteService {
+    PagedResult<Route> findAllRoutes(int pageNo, int pageSize, String sortBy, String sortDir);
 
-    @Autowired private RouteRepository routeRepository;
+    Optional<Route> findRouteById(Long id);
 
-    public PagedResult<Route> findAllRoutes(
-            int pageNo, int pageSize, String sortBy, String sortDir) {
-        Sort sort =
-                sortDir.equalsIgnoreCase(Sort.Direction.ASC.name())
-                        ? Sort.by(sortBy).ascending()
-                        : Sort.by(sortBy).descending();
+    Route saveRoute(Route route);
 
-        // Create Pageable instance
-        Pageable pageable = PageRequest.of(pageNo, pageSize, sort);
-        Page<Route> routesPage = routeRepository.findAll(pageable);
-
-        return new PagedResult<>(routesPage);
-    }
-
-    public Optional<Route> findRouteById(Long id) {
-        return routeRepository.findById(id);
-    }
-
-    public Route saveRoute(Route route) {
-        return routeRepository.save(route);
-    }
-
-    public void deleteRouteById(Long id) {
-        routeRepository.deleteById(id);
-    }
+    void deleteRouteById(Long id);
 }
